Revamp premium calculation, remove Kafka and optimize service logic

+++ business-services/property-quote-service/src/main/java/edu/hcmute/dto/CoverageDto.java
@@ -6,6 +6,6 @@ public record CoverageDto(
         Integer id,
         CoverageCode code,
         Long limit,
-        Long deductible
+        Double deductible
 ) {
 }
